类型修改 Integer -> Long

+++ jeecg-boot/jeecg-boot-module-system/src/main/java/org/jeecg/modules/game/entity/GameVirtualOrder.java
@@ -29,7 +29,7 @@ public class GameVirtualOrder implements Serializable {
      * 自增主键
      */
     @TableId(type = IdType.AUTO)
-    private java.lang.Integer id;
+    private java.lang.Long id;
 
     /**
      * 区服id
